Remove Btcruby namespace
- All btc address creation tasks use btcruby.
- No logic change
