--- app/services/btc/addresses/creation/btcruby/get_address_index.rb
@@ -1,18 +1,0 @@
-module Btc
-  module Addresses
-    module Creation
-      module Btcruby
-        class GetAddressIndex
-
-          extend LightService::Action
-          promises :address_index
-
-          executed do |c|
-            c.address_index = Address.btc.count
-          end
-
-        end
-      end
-    end
-  end
-end
